# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Needify", "description": "This is a service for Needify app", "version": "1.0.0"}, "paths": {"/v1/users/list": {"get": {"tags": ["users"], "summary": "Get Users", "operationId": "get_users_v1_users_list_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/UserSchema"}, "type": "array", "title": "Response Get Users V1 Users List Get"}}}}}}}, "/v1/users/{telegram_id}": {"get": {"tags": ["users"], "summary": "Get User", "operationId": "get_user_v1_users__telegram_id__get", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Telegram Id"}, "name": "telegram_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["users"], "summary": "Update User", "operationId": "update_user_v1_users__telegram_id__put", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Telegram Id"}, "name": "telegram_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateUserSchema"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["users"], "summary": "Delete User", "operationId": "delete_user_v1_users__telegram_id__delete", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Telegram Id"}, "name": "telegram_id", "in": "path"}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/users/": {"post": {"tags": ["users"], "summary": "Create User", "operationId": "create_user_v1_users__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateUserSchema"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/users/wallet/add": {"post": {"tags": ["users"], "summary": "Add Wallet User", "operationId": "add_wallet_user_v1_users_wallet_add_post", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Telegram Id"}, "name": "telegram_id", "in": "query"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserWeb3WalletSchema"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/currency/": {"get": {"tags": ["currency"], "summary": "Get List Currencies", "operationId": "get_list_currencies_v1_currency__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/CurrencySchema"}, "type": "array", "title": "Response Get List Currencies V1 Currency  Get"}}}}}}}, "/v1/currency/native-currency": {"get": {"tags": ["currency"], "summary": "Get Native Currency", "operationId": "get_native_currency_v1_currency_native_currency_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CurrencySchema"}}}}}}}, "/v1/task/": {"get": {"tags": ["task"], "summary": "Get List Tasks", "operationId": "get_list_tasks_v1_task__get", "parameters": [{"required": false, "schema": {"type": "string", "title": "Category"}, "name": "category", "in": "query"}, {"required": false, "schema": {"allOf": [{"$ref": "#/components/schemas/TaskStatusEnum"}], "default": "PUBLISHED"}, "name": "status_", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskSchema", "items": {"$ref": "#/components/schemas/TaskSchema"}, "type": "array", "title": "Response Get List Tasks V1 Task  Get"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["task"], "summary": "Create Task", "operationId": "create_task_v1_task__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PreCreateTaskSchema"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskSchema"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/task/{task_id}": {"get": {"tags": ["task"], "summary": "Get Task", "operationId": "get_task_v1_task__task_id__get", "parameters": [{"required": true, "schema": {"type": "integer", "title": "Task Id"}, "name": "task_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskSchema"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/task/{user_id}/tasks": {"get": {"tags": ["task"], "summary": "Get Tasks By User Id", "operationId": "get_tasks_by_user_id_v1_task__user_id__tasks_get", "parameters": [{"required": true, "schema": {"type": "integer", "title": "User Id"}, "name": "user_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserHistoryResponseSchema"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/job-offer/message/deploy": {"post": {"tags": ["job-offer"], "summary": "Get Deploy Job Offer Message", "operationId": "get_deploy_job_offer_message_v1_job_offer_message_deploy_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobOfferMessageSchema"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobOfferMessageDeployResponseSchema"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/job-offer/message/get-job": {"post": {"tags": ["job-offer"], "summary": "Get Job Offer Message", "operationId": "get_job_offer_message_v1_job_offer_message_get_job_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/GetJob"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobOfferMessageResponseSchema"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/job-offer/message/choose-doer": {"post": {"tags": ["job-offer"], "summary": "Choose Doer Job Offer Message", "operationId": "choose_doer_job_offer_message_v1_job_offer_message_choose_doer_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChooseDoerSchema"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobOfferMessageResponseSchema"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/job-offer/message/complete": {"post": {"tags": ["job-offer"], "summary": "Complete Job Offer Message", "operationId": "complete_job_offer_message_v1_job_offer_message_complete_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CompleteJob"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobOfferMessageResponseSchema"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/job-offer/message/confirm": {"post": {"tags": ["job-offer"], "summary": "Complete Job Offer Message", "operationId": "complete_job_offer_message_v1_job_offer_message_confirm_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfirmJob"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobOfferMessageResponseSchema"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/job-offer/message/revoke": {"post": {"tags": ["job-offer"], "summary": "Complete Job Offer Message", "operationId": "complete_job_offer_message_v1_job_offer_message_revoke_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RevokeJob"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobOfferMessageResponseSchema"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseErrorResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/category/": {"get": {"tags": ["category"], "summary": "Get Categories", "operationId": "get_categories_v1_category__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/CategorySchema"}, "type": "array", "title": "Response Get Categories V1 Category  Get"}}}}}}, "post": {"tags": ["category"], "summary": "Create Categories", "operationId": "create_categories_v1_category__post", "requestBody": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/CategorySchema"}, "type": "array", "title": "Categories"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/CategorySchema"}, "type": "array", "title": "Response Create Categories V1 Category  Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/v1/ton-connect/test": {"post": {"tags": ["ton-connect"], "summary": "Get Last Masterchain Block", "operationId": "get_last_masterchain_block_v1_ton_connect_test_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConnectDepositSchema"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"BaseErrorResponse": {"properties": {"error": {"$ref": "#/components/schemas/ErrorSchema"}}, "type": "object", "required": ["error"], "title": "BaseErrorResponse"}, "CategorySchema": {"properties": {"title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}}, "type": "object", "required": ["title", "description"], "title": "CategorySchema"}, "ChooseDoerSchema": {"properties": {"task_id": {"type": "integer", "title": "Task Id"}, "action_by_user": {"type": "integer", "title": "Action By User"}, "doer": {"type": "string", "title": "Doer"}}, "type": "object", "required": ["task_id", "action_by_user", "doer"], "title": "ChooseDoerSchema"}, "CompleteJob": {"properties": {"task_id": {"type": "integer", "title": "Task Id"}, "action_by_user": {"type": "integer", "title": "Action By User"}}, "type": "object", "required": ["task_id", "action_by_user"], "title": "CompleteJob"}, "ConfirmJob": {"properties": {"task_id": {"type": "integer", "title": "Task Id"}, "action_by_user": {"type": "integer", "title": "Action By User"}, "mark": {"type": "integer", "title": "Mark"}, "review": {"type": "string", "title": "Review"}}, "type": "object", "required": ["task_id", "action_by_user"], "title": "ConfirmJob"}, "ConnectDepositSchema": {"properties": {"action_by_user_id": {"type": "integer", "title": "Action By User Id"}, "jetton_wallet_address": {"type": "string", "title": "Jetton Wallet Address"}, "recipient_address": {"type": "string", "title": "Recipient Address"}, "transfer_fee": {"type": "number", "title": "Transfer Fee"}, "jettons_amount": {"type": "number", "title": "Jettons Amount"}, "wallet_name": {"type": "string", "title": "Wallet Name"}}, "type": "object", "required": ["action_by_user_id", "jetton_wallet_address", "recipient_address", "transfer_fee", "jettons_amount", "wallet_name"], "title": "ConnectDepositSchema"}, "CreateUserSchema": {"properties": {"telegram_id": {"type": "integer", "title": "Telegram Id"}, "username": {"type": "string", "title": "Username"}, "first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "image": {"type": "string", "title": "Image"}, "permissions": {"items": {"type": "string"}, "type": "array", "title": "Permissions"}, "disabled": {"type": "boolean", "title": "Disabled", "default": false}}, "type": "object", "required": ["telegram_id"], "title": "CreateUserSchema"}, "CurrencySchema": {"properties": {"jetton_master_address": {"type": "string", "title": "Jetton Master Address"}, "decimals": {"type": "integer", "title": "Decimals"}, "symbol": {"type": "string", "title": "Symbol"}, "name": {"type": "string", "title": "Name"}, "is_active": {"type": "boolean", "title": "Is Active"}}, "type": "object", "required": ["jetton_master_address", "decimals", "symbol", "name", "is_active"], "title": "CurrencySchema"}, "ErrorSchema": {"properties": {"name": {"type": "string", "title": "Name"}, "code": {"type": "integer", "title": "Code"}, "description": {"type": "string", "title": "Description"}, "meta": {"type": "object", "title": "Meta"}}, "type": "object", "required": ["name", "code", "description", "meta"], "title": "ErrorSchema"}, "GetJob": {"properties": {"task_id": {"type": "integer", "title": "Task Id"}, "action_by_user": {"type": "integer", "title": "Action By User"}}, "type": "object", "required": ["task_id", "action_by_user"], "title": "GetJob"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "JobOfferMessageDeployResponseSchema": {"properties": {"valid_until": {"type": "integer", "title": "Valid Until"}, "messages": {"items": {"$ref": "#/components/schemas/TONConnectMessageSchema"}, "type": "array", "title": "Messages"}, "stateInit": {"type": "string", "title": "Stateinit"}}, "type": "object", "required": ["valid_until", "messages", "stateInit"], "title": "JobOfferMessageDeployResponseSchema"}, "JobOfferMessageResponseSchema": {"properties": {"valid_until": {"type": "integer", "title": "Valid Until"}, "messages": {"items": {"$ref": "#/components/schemas/TONConnectMessageSchema"}, "type": "array", "title": "Messages"}}, "type": "object", "required": ["valid_until", "messages"], "title": "JobOfferMessageResponseSchema"}, "JobOfferMessageSchema": {"properties": {"task_id": {"type": "integer", "title": "Task Id"}, "action_by_user": {"type": "integer", "title": "Action By User"}}, "type": "object", "required": ["task_id", "action_by_user"], "title": "JobOfferMessageSchema"}, "JobOfferSchema": {"properties": {"job_offer_address": {"type": "string", "title": "Job Offer Address"}, "jetton_master_address": {"type": "string", "title": "Jetton Master Address"}, "jetton_native_address": {"type": "string", "title": "Jetton Native Address"}, "state": {"type": "string", "title": "State"}, "stateInit": {"type": "string", "title": "Stateinit"}, "owner": {"type": "string", "title": "Owner"}, "doer": {"type": "string", "title": "Doer"}, "vacancies": {"items": {"$ref": "#/components/schemas/VacancySchema"}, "type": "array", "title": "Vacancies"}, "mark": {"type": "integer", "title": "Mark"}, "review": {"type": "string", "title": "Review"}}, "type": "object", "required": ["job_offer_address", "jetton_master_address", "jetton_native_address", "state", "stateInit", "owner", "vacancies"], "title": "JobOfferSchema"}, "PreCreateTaskSchema": {"properties": {"title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "category": {"type": "string", "title": "Category"}, "images": {"items": {"type": "string"}, "type": "array", "title": "Images", "description": "Image in string format"}, "price": {"type": "number", "exclusiveMinimum": 0.0, "title": "Price"}, "currency": {"type": "string", "title": "Currency", "description": "Currency symbol"}, "poster_id": {"type": "integer", "exclusiveMinimum": 0.0, "title": "Poster Id", "description": "Customer telegram ID"}, "poster_address": {"type": "string", "title": "Poster Address", "description": "Customer wallet jetton_master_address"}, "deadline": {"type": "string", "format": "date-time", "title": "Deadline"}}, "type": "object", "required": ["title", "description", "category", "images", "price", "currency", "poster_id", "poster_address", "deadline"], "title": "PreCreateTaskSchema"}, "RevokeJob": {"properties": {"task_id": {"type": "integer", "title": "Task Id"}, "action_by_user": {"type": "integer", "title": "Action By User"}}, "type": "object", "required": ["task_id", "action_by_user"], "title": "RevokeJob"}, "TONConnectMessageSchema": {"properties": {"address": {"type": "string", "title": "Address", "description": "TON/JettonWallet/etc jetton_master_address"}, "amount": {"type": "string", "title": "Amount", "description": "String amount in nanoTON"}, "payload": {"type": "string", "title": "Payload", "description": "Base64 encoded string"}}, "type": "object", "required": ["address", "amount", "payload"], "title": "TONConnectMessageSchema"}, "TaskSchema": {"properties": {"task_id": {"type": "integer", "title": "Task Id"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "images": {"items": {"type": "string"}, "type": "array", "title": "Images", "default": []}, "category": {"type": "string", "title": "Category"}, "price": {"type": "number", "exclusiveMinimum": 0.0, "title": "Price"}, "currency": {"type": "string", "title": "Currency"}, "status": {"allOf": [{"$ref": "#/components/schemas/TaskStatusEnum"}], "default": "PRE_CREATED"}, "poster_id": {"type": "integer", "exclusiveMinimum": 0.0, "title": "Poster Id"}, "poster_address": {"type": "string", "title": "Poster Address"}, "doer_id": {"type": "integer", "title": "Doer Id"}, "doer_address": {"type": "string", "title": "Doer Address"}, "job_offer": {"$ref": "#/components/schemas/JobOfferSchema"}, "deadline": {"type": "string", "format": "date-time", "title": "Deadline"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["task_id", "title", "description", "category", "price", "currency", "poster_id", "poster_address", "deadline", "created_at"], "title": "TaskSchema"}, "TaskStatusEnum": {"type": "string", "enum": ["PRE_CREATED", "DEPLOYING", "PUBLISHED", "CLOSED", "IN_PROGRESS", "COMPLETED", "CONFIRMED"], "title": "TaskStatusEnum", "description": "An enumeration."}, "UpdateUserSchema": {"properties": {"first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "image": {"type": "string", "title": "Image"}}, "type": "object", "title": "UpdateUserSchema"}, "UserHistoryResponseSchema": {"properties": {"published_tasks": {"items": {"$ref": "#/components/schemas/TaskSchema"}, "type": "array", "title": "Published Tasks"}, "picked_up_tasks": {"items": {"$ref": "#/components/schemas/TaskSchema"}, "type": "array", "title": "Picked Up Tasks"}, "completed_tasks": {"items": {"$ref": "#/components/schemas/TaskSchema"}, "type": "array", "title": "Completed Tasks"}}, "type": "object", "required": ["published_tasks", "picked_up_tasks", "completed_tasks"], "title": "UserHistoryResponseSchema"}, "UserSchema": {"properties": {"_id": {"type": "string", "title": " Id"}, "telegram_id": {"type": "integer", "title": "Telegram Id"}, "first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "username": {"type": "string", "title": "Username"}, "image": {"type": "string", "title": "Image"}, "permissions": {"items": {"type": "string"}, "type": "array", "title": "Permissions"}, "web3_wallet": {"$ref": "#/components/schemas/UserWeb3WalletSchema"}, "disabled": {"type": "boolean", "title": "Disabled"}, "created": {"type": "string", "format": "date-time", "title": "Created"}, "updated": {"type": "string", "format": "date-time", "title": "Updated"}}, "type": "object", "required": ["_id", "telegram_id", "created"], "title": "users"}, "UserWeb3WalletSchema": {"properties": {"address": {"type": "string", "title": "Address"}}, "type": "object", "required": ["address"], "title": "UserWeb3WalletSchema"}, "VacancySchema": {"properties": {"doer": {"type": "string", "title": "Doer"}, "telegram_id": {"type": "integer", "title": "Telegram Id"}}, "type": "object", "required": ["doer", "telegram_id"], "title": "VacancySchema"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}